Rename the package
The word "framework" and phrase "the framework" (referring to this
entire software package) have been converted to appropriate usages of
"Tinker Engine" (referring to the full package), "tinker" (referring to
the name of the executable), or left as "framework" in its generic sense
of a software system that other things can hook into.

This includes renaming the Python package directory to `tinker-engine`,
the console script associated with the executable version to `tinker`,
and updating other references in code, comments, and elsewhere as well.

This leaves the documentation and Dockerfile intact, as those will be
updated in future efforts.